Former-commit-id: c2c508d49534d9eaf83e6af136f86c2eaade8256 [formerly 8953fcf4afd84ec0ca8d80e697c0a3c8f378b72c] [formerly c2c508d49534d9eaf83e6af136f86c2eaade8256 [formerly 8953fcf4afd84ec0ca8d80e697c0a3c8f378b72c] [formerly 8a6844e5a3b4f0cacc26df3aae65bda8b6b7cf2a [formerly 93c0e07b71d33560678eb3781465e20e9222642d]]]
Former-commit-id: 8a6844e5a3b4f0cacc26df3aae65bda8b6b7cf2a
Former-commit-id: a9395e64abcaa7c1a5f15e232efea64564d8f49c [formerly 8208937771fcac0a64e1f5759502b26562d63441]
Former-commit-id: d57d70be63fb3fa6f6fac1d838ac168bc53f2abb

+++ pythonPackages/pypies/pypies/impl/H5pyDataStore.py
@@ -322,9 +322,16 @@ class H5pyDataStore(IDataStore.IDataStore):
         try:
             locs = request.getLocations()
             for dataset in locs:
-                ds = self.__getNode(f, None, dataset)
-                grp = ds.parent
-                grp.id.unlink(ds.name)
+                ds = None
+
+                try :
+                    ds = self.__getNode(f, None, dataset)
+                except Exception, e:
+                    logger.warn('Unable to find uri [' + str(dataset) + '] in file [' + str(fn) + '] to delete: ' + IDataStore._exc())
+
+                if ds:
+                    grp = ds.parent
+                    grp.id.unlink(ds.name)
 
         finally:
             # check if file has any remaining data sets
